+++ cucumber-expressions/go/parameter_type.go
@@ -24,7 +24,7 @@ func CheckParameterTypeName(typeName string) error {
 	unescapedTypeName := UNESCAPE_REGEXP.ReplaceAllString(typeName, "$2")
 	if ILLEGAL_PARAMETER_NAME_REGEXP.MatchString(typeName) {
 		c := ILLEGAL_PARAMETER_NAME_REGEXP.FindStringSubmatch(typeName)[0]
-		return errors.New(fmt.Sprintf("Illegal character '%s' in parameter name {%s}", c, unescapedTypeName))
+		return fmt.Errorf("Illegal character '%s' in parameter name {%s}", c, unescapedTypeName)
 	}
 	return nil
 }